Add Product-in-cart schema & order schema and refactor old schemas
Order schema contain the typical structure an order like cost, transaction
ID, address, user.
Product-in-cart schema is different from product schema because once the
products have been placed in the cart, the structure changes because now
we have to account for the quantity as well.
Minor refactoring also done to old schemas