Add new field in the order model
This update adds a new field in the order schema called "status". This
field has the type of enum and can store information about the status
of the order for tracking purposes.
Also changed the type of "updated" field to "Date" because it now shows
the date of the last status update.

Signed-off-by: PushpakB3096 <rtpushpak@gmail.com>